+++ src/js/presenters/controls/UpdateModalPresenter.ts
@@ -1,10 +1,10 @@
 
 import ModalPresenter from "~/js/core/presenters/ModalPresenter";
-import Presenter, { on, restore } from "~/js/core/base/Presenter";
+import {on} from "~/js/core/base/Presenter";
 import ConnectionModel, { ConnectionStatusEvent } from "~/js/models/common/ConnectionModel";
 import ServerModel from "~/js/models/common/ServerModel";
 import i18next from "i18next";
-import { ModalAction } from "~/js/core/base/view/Nest";
+import {ModalAction} from "~/js/core/views/modal/ModalView";
 
 // passed by DefinePlugin in Webpack config
 declare const __VERSION__: string;